// src/models/walletswallet-idpaymentfees-payments.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Cardano Wallet Backend API
 * <p align=\"right\"><img style=\"position: relative; top: -10em; margin-bottom: -12em;\" width=\"20%\" src=\"https://cardanodocs.com/img/cardano.png\"></img></p> 
 *
 * OpenAPI spec version: 2021.3.4
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

import { WalletsAssetsAvailable, WalletswalletIdpaymentfeesAmount } from ".";

/**
 * 
 * @export
 * @interface WalletswalletIdpaymentfeesPayments
 */
export interface WalletswalletIdpaymentfeesPayments {
    /**
     * 
     * @type {string}
     * @memberof WalletswalletIdpaymentfeesPayments
     */
    address: string;
    /**
     * 
     * @type {WalletswalletIdpaymentfeesAmount}
     * @memberof WalletswalletIdpaymentfeesPayments
     */
    amount: WalletswalletIdpaymentfeesAmount;
    /**
     * A flat list of assets.
     * @type {Array&lt;WalletsAssetsAvailable&gt;}
     * @memberof WalletswalletIdpaymentfeesPayments
     */
    assets?: Array<WalletsAssetsAvailable>;

    /**
     * 
     * @type {any}
     * @memberof WalletswalletIdpaymentfeesPayments
     */
    data?: { hash? :string, asHash?: string, inline: string };


    /**
     * 
     * @type {string}
     * @memberof WalletswalletIdpaymentfeesPayments
     */
    script_ref?: string;
}